Reorganize services/direct into domain folders
Almost nothing in services/direct was direct-mode-specific: the custody
layer is consumed by the fireblocks/dfns integrations, deposits and the
plan-approval options; the mapping domain by finp2p-contract and both
custody modes; the omnibus delegate is omnibus-mode outright. Split by
domain instead of one mode folder:

- services/custody/    provider, registry, gas-station, wallet-activation,
                       wallet-resolver, health (mode-agnostic wallet layer)
- services/accounts/   account-mapping, mapping-validator, mapping-config
- services/operations/ operation-status (was common.ts), operation-context
- services/assets/     store (was helpers.ts, getAssetFromDb)
- services/omnibus/    omnibus-delegate (moved out of direct)
- services/direct/     token-service only — the actual direct-mode service

All ~20 import sites updated (app, integrations, finp2p-contract,
plan-approval options, tests); no re-export shims left behind.
