+++ src/integrations/deposits/ota-deposit/plugin.ts
@@ -14,7 +14,8 @@ import {
   failedDepositOperation,
 } from "@owneraio/finp2p-nodejs-skeleton-adapter";
 import { FinP2PClient } from "@owneraio/finp2p-client";
-import { AssetStore, CustodyProvider } from "../../../services/direct";
+import { CustodyProvider } from "../../../services/custody";
+import { AssetStore } from "../../../services/accounts";
 import { DepositTargetResolver } from "../types";
 import { BalanceWatcher } from "./balance-watcher";
 import { OtaResult } from "./models";
