+++ src/integrations/fireblocks/provider.ts
@@ -2,7 +2,7 @@ import { FireblocksSDK } from 'fireblocks-sdk';
 import axios from 'axios';
 import { createFireblocksEthersProvider, FireblocksAppConfig } from './config';
 import { createVaultManagementFunctions } from '../../vaults';
-import { CustodyProvider, CustodyWallet, GasStation } from '../../services/direct';
+import { CustodyProvider, CustodyWallet, GasStation } from '../../services/custody';
 import { FireblocksRawSigner } from './raw-signer';
 
 export class FireblocksCustodyProvider implements CustodyProvider {
