+++ src/services/direct/index.ts
@@ -1,12 +1,1 @@
-export * from './custody-provider'
-export * from './gas-station'
-export * from './custody-registry'
-export * from './account-mapping'
 export * from './token-service'
-export * from './common'
-export * from './health'
-export * from './omnibus-delegate'
-export * from './mapping-validator'
-export * from './mapping-config'
-export * from './wallet-resolver'
-export * from './wallet-activation'
