+++ src/services/direct/token-service.ts
@@ -7,13 +7,13 @@ import {
 import winston from 'winston';
 import { parseUnits } from "ethers";
 import { TokenOperationResult } from '@owneraio/finp2p-ethereum-adapter-contract';
-import { CustodyProvider, CustodyWallet } from './custody-provider';
-import { AccountMappingService, AssetStore } from './account-mapping';
-import { getAssetFromDb } from './helpers';
+import { CustodyProvider, CustodyWallet } from '../custody/custody-provider';
+import { AccountMappingService, AssetStore } from '../accounts/account-mapping';
+import { getAssetFromDb } from '../assets/store';
 import { tokenStandardRegistry } from '../../integrations/token-standards/registry';
 import { TokenStandardName as ERC20_TOKEN_STANDARD, DEFAULT_NEW_ERC20_DECIMALS } from '@owneraio/finp2p-ethereum-erc20-plugin';
 import { ERC20Contract } from '@owneraio/finp2p-contracts';
-import { buildOperationContext } from './operation-context';
+import { buildOperationContext } from '../operations/operation-context';
 
 function resultToReceipt(
   result: TokenOperationResult, ast: Asset, operationType: OperationType, quantity: string,
