+++ src/services/finp2p-contract/mapping.ts
@@ -1,6 +1,6 @@
 import { AccountMappingService, AccountMapping, ReceiptOperation, Asset, ExecutionContext } from '@owneraio/finp2p-nodejs-skeleton-adapter';
 import { FinP2PContract, ReceiptOperation as ContractReceiptOperation } from '@owneraio/finp2p-contracts';
-import { FIELD_LEDGER_ACCOUNT_ID } from '../direct/mapping-validator';
+import { FIELD_LEDGER_ACCOUNT_ID } from '../accounts/mapping-validator';
 
 function mapAccount(acc: { finId: string; account?: string } | undefined) {
   if (!acc || !acc.finId) return undefined;
